Extract Failure into failure.dart
