+++ lib/src/sdk/no_updates_failure.dart
@@ -1,4 +1,4 @@
-import 'errors.dart' show Failure;
+import 'failure.dart' show Failure;
 
 class NoUpdatesFailure extends Failure {
   NoUpdatesFailure() : super(message: 'No updates');
